feat/lrm: refactor mixinMessage to mixinWebsocket, add socket status indicator to the footer
