+++ web/lrm/client/pages/demo.vue
@@ -120,10 +120,10 @@
 import { toRef } from 'vue'
 import { toast } from 'vue3-toastify'
 import { consola } from 'consola'
-import mixinMessage from '~/mixins/mixinMessage'
+import mixinWebsocket from '~/mixins/mixinWebsocket'
 import { REPOSITORY_URL } from '@/constants'
 import { useMainStore } from '~/store'
-import { buildMessage, sendMessage } from '~/composables/messageUtils.js'
+import { buildMessage, sendMessage, isOut, getMessageType, getCaseId } from '~/composables/messageUtils.js'
 
 function submit (form) {
   try {
@@ -144,7 +144,7 @@ useHead({
 <script>
 export default {
   name: 'Demo',
-  mixins: [mixinMessage],
+  mixins: [mixinWebsocket],
   beforeRouteEnter (to, from) {
     // Redirect to parent if we're not authenticated
     if (!useMainStore().isAuthenticated) {
@@ -199,16 +199,16 @@ export default {
     clientMessages () {
       return this.store.messages.filter(
         message => (
-          (this.isOut(message.direction) && message.body.senderID === this.store.user.clientId) ||
-          (!this.isOut(message.direction) && message.routingKey.startsWith(this.store.user.clientId))
+          (isOut(message.direction) && message.body.senderID === this.store.user.clientId) ||
+          (!isOut(message.direction) && message.routingKey.startsWith(this.store.user.clientId))
         )
       )
     },
     showableMessages () {
-      return this.store.showSentMessages ? this.clientMessages : this.clientMessages?.filter(message => !this.isOut(message.direction))
+      return this.store.showSentMessages ? this.clientMessages : this.clientMessages?.filter(message => !isOut(message.direction))
     },
     selectedTypeMessages () {
-      return this.showableMessages.filter(message => this.getMessageType(message) === this.selectedMessageType)
+      return this.showableMessages.filter(message => getMessageType(message) === this.selectedMessageType)
     },
     selectedVhost () {
       return this.store.selectedVhost
@@ -218,14 +218,14 @@ export default {
         return this.selectedTypeMessages
       }
       return this.selectedTypeMessages.filter(
-        message => this.selectedCaseIds.includes(this.getCaseId(message, true))
+        message => this.selectedCaseIds.includes(getCaseId(message, true))
       )
     },
     messagesSentCount () {
-      return this.clientMessages.filter(message => this.isOut(message.direction)).length
+      return this.clientMessages.filter(message => isOut(message.direction)).length
     },
     caseIds () {
-      return [...new Set(this.selectedTypeMessages.map(m => this.getCaseId(m, true)))]
+      return [...new Set(this.selectedTypeMessages.map(m => getCaseId(m, true)))]
     }
   },
   watch: {
@@ -262,7 +262,7 @@ export default {
     },
     typeMessages (type) {
       return this.showableMessages.filter(
-        message => this.getMessageType(message) === type
+        message => getMessageType(message) === type
       )
     },
