+++ web/lrm/client/pages/test/case.vue
@@ -181,7 +181,7 @@
 <script setup>
 import { onMounted, toRefs } from 'vue'
 import jsonpath from 'jsonpath'
-import mixinMessage from '~/mixins/mixinMessage'
+import mixinWebsocket from '~/mixins/mixinWebsocket'
 import { useMainStore } from '~/store'
 import { REPOSITORY_URL } from '@/constants'
 import { isOut, getCaseId, getMessageType, setCaseId, buildMessage, sendMessage } from '~/composables/messageUtils.js'
@@ -543,7 +543,7 @@ watch(selectedTypeCaseMessages, (newMessages) => {
 <script>
 export default {
   name: 'Testcase',
-  mixins: [mixinMessage],
+  mixins: [mixinWebsocket],
   beforeRouteEnter (to, from) {
     if (!useMainStore().isAuthenticated) {
       return { name: 'index' }
